Move apps/api to apps/rest_api.
