+++ blockstore/apps/rest_api/urls.py
@@ -9,5 +9,5 @@ from django.conf.urls import url, include
 app_name = 'blockstore'
 
 urlpatterns = [
-    url(r'^v1/', include('blockstore.apps.api.v1.urls', namespace='v1')),
+    url(r'^v1/', include('blockstore.apps.rest_api.v1.urls', namespace='v1')),
 ]
